+++ ql/src/java/org/apache/hadoop/hive/ql/optimizer/graph/OperatorGraph.java
@@ -20,8 +20,6 @@
 package org.apache.hadoop.hive.ql.optimizer.graph;
 
 import java.io.File;
-import java.io.PrintWriter;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashSet;
@@ -29,7 +27,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import org.apache.curator.shaded.com.google.common.base.Joiner;
 import org.apache.hadoop.hive.ql.exec.AppMasterEventOperator;
 import org.apache.hadoop.hive.ql.exec.Operator;
 import org.apache.hadoop.hive.ql.exec.ReduceSinkOperator;
@@ -37,10 +34,6 @@ import org.apache.hadoop.hive.ql.exec.TableScanOperator;
 import org.apache.hadoop.hive.ql.parse.ParseContext;
 import org.apache.hadoop.hive.ql.parse.SemiJoinBranchInfo;
 import org.apache.hadoop.hive.ql.plan.DynamicPruningEventDesc;
-import org.apache.hadoop.hive.ql.plan.JoinDesc;
-import org.apache.hadoop.hive.ql.plan.OperatorDesc;
-import org.apache.hadoop.hive.ql.plan.TableScanDesc;
-
 import com.google.common.collect.Sets;
 
 /**
@@ -158,80 +151,7 @@ public class OperatorGraph {
   }
 
   public void toDot(File outFile) throws Exception {
-    PrintWriter writer = new PrintWriter(outFile);
-    writer.println("digraph G");
-    writer.println("{");
-    HashSet<Cluster> clusters = new HashSet<>(nodeCluster.values());
-    int idx = 0;
-    for (Cluster cluster : clusters) {
-      idx++;
-      writer.printf("subgraph cluster_%d {", idx);
-      for (Operator<?> member : cluster.members) {
-        writer.printf("%s;", nodeName(member));
-      }
-      writer.printf("label = \"cluster %d\";", idx);
-      writer.printf("}");
-    }
-    Set<Operator<?>> nodes = g.nodes();
-    for (Operator<?> n : nodes) {
-      writer.printf("%s[shape=record,label=\"%s\",%s];", nodeName(n), nodeLabel(n), style(n));
-      Set<Operator<?>> succ = g.successors(n);
-      for (Operator<?> s : succ) {
-        writer.printf("%s->%s;", nodeName(n), nodeName(s));
-      }
-    }
-
-    writer.println("}");
-    writer.close();
-  }
-
-
-  private String style(Operator<?> n) {
-    String fillColor = "white";
-    OperatorDesc c = n.getConf();
-    if (n instanceof TableScanOperator) {
-      fillColor = "#ccffcc";
-    }
-    if (c instanceof JoinDesc) {
-      fillColor = "#ffcccc";
-    }
-    return String.format("style=filled,fillcolor=\"%s\"", fillColor);
-  }
-
-  private String nodeLabel(Operator<?> n) {
-    List<String> rows = new ArrayList<String>();
-
-    rows.add(nodeName0(n));
-    if ((n instanceof TableScanOperator)) {
-      TableScanOperator ts = (TableScanOperator) n;
-      TableScanDesc conf = ts.getConf();
-      rows.add(vBox(conf.getTableName(), conf.getAlias()));
-    }
-    return vBox(rows);
-  }
-
-  private String hBox(List<String> rows) {
-    return "" + Joiner.on("|").join(rows) + "";
-  }
-
-  private String hBox(String... rows) {
-    return "" + Joiner.on("|").join(rows) + "";
-  }
-  private String vBox(List<String> rows) {
-    return "{ " + hBox(rows) + "}";
-  }
-
-  private String vBox(String... strings) {
-    return "{ " + hBox(strings) + "}";
-  }
-
-
-  private String nodeName(Operator<?> member) {
-    return String.format("\"%s\"", member);
-  }
-
-  private String nodeName0(Operator<?> member) {
-    return member.toString();
+    new DotExporter(this).write(outFile);
   }
 
   public boolean mayMerge(Operator<?> opA, Operator<?> opB) {
